backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
av==14.0.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
jiter==0.8.2
joblib==1.4.2
livekit==0.19.1
livekit-agents==0.12.7
livekit-api==0.8.1
livekit-plugins-deepgram==0.6.16
livekit-plugins-openai==0.10.14
livekit-plugins-silero==0.7.4
livekit-protocol==0.8.0
llama-index==0.10.18
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-clip==0.1.5
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-groq==0.1.3
llama-index-llms-openai==0.1.31
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-lancedb==0.1.7
llama-parse==0.4.9
llamaindex-py-client==0.1.19
marshmallow==3.24.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
onnxruntime==1.20.1
openai==1.59.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.8
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pypdf==4.0.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.3
tenacity==8.2.3
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.24.0
wrapt==1.17.0
yarl==1.18.3